Module Store.Impl

type 'a t =

| Hash_keyed : (module Irmin.S with type t = 'a) -> 'a t

| Generic_keyed : (module Irmin.Generic_key.S with type t = 'a) -> 'a t

The type of implementations of an Irmin store.

Stores can be either keyed by hashes or by some other abstract type. In the latter case, the store implementation cannot be used to build HTTP / GraphQL servers using irmin-unix. This limitation may be lifted in a future version of irmin-unix.

val generic_keyed : 'a t -> (module Irmin.Generic_key.S with type t = 'a)

val hash_keyed : 'a t -> (module Irmin.S with type t = 'a) option